requirements.txt
alembic==1.4.2
bcrypt==3.1.7
branca==0.4.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
folium==0.11.0
geographiclib==1.50
geopy==1.22.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
pycparser==2.20
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
-e git+https://github.com/AlexandrBashkirtsev/gascalc.git@0f293e18f01752cce0c8039923f22f100143bece#egg=utils&subdirectory=gascalc/plugins/utils
Werkzeug==1.0.1
WTForms==2.3.1
